# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stripe Service API",
    "description": "Payment processing service built on Stripe. Handles checkout sessions, payment intents, webhook processing for payment events, and integrates with runs-service for cost tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://stripe.mcpfactory.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:3011",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
      "CreateCheckoutSessionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "paymentId": {
            "type": "string",
            "format": "uuid",
            "description": "Internal payment record ID"
          },
          "sessionId": {
            "type": "string",
            "description": "Stripe Checkout Session ID"
          },
          "url": {
            "type": "string",
            "description": "Stripe Checkout URL"
          }
        },
        "required": [
          "success",
          "paymentId"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "nullable": true,
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateCheckoutSessionRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "description": "Brand ID"
          },
          "campaignId": {
            "type": "string",
            "description": "Campaign ID"
          },
          "lineItems": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "priceId": {
                  "type": "string",
                  "description": "Stripe Price ID"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true,
                  "description": "Quantity"
                }
              },
              "required": [
                "priceId",
                "quantity"
              ]
            },
            "minItems": 1,
            "description": "Line items for checkout"
          },
          "successUrl": {
            "type": "string",
            "format": "uri",
            "description": "Redirect URL on success"
          },
          "cancelUrl": {
            "type": "string",
            "format": "uri",
            "description": "Redirect URL on cancel"
          },
          "customerId": {
            "type": "string",
            "description": "Stripe Customer ID"
          },
          "customerEmail": {
            "type": "string",
            "format": "email",
            "description": "Customer email"
          },
          "mode": {
            "type": "string",
            "enum": [
              "payment",
              "subscription"
            ],
            "default": "payment",
            "description": "Checkout mode"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom metadata"
          },
          "discounts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "coupon": {
                  "type": "string",
                  "description": "Stripe Coupon ID"
                },
                "promotionCode": {
                  "type": "string",
                  "description": "Stripe Promotion Code ID"
                }
              }
            },
            "description": "Discounts to apply to the checkout session"
          }
        },
        "required": [
          "lineItems",
          "successUrl",
          "cancelUrl"
        ]
      },
      "CreatePaymentIntentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "paymentId": {
            "type": "string",
            "format": "uuid",
            "description": "Internal payment record ID"
          },
          "paymentIntentId": {
            "type": "string",
            "description": "Stripe Payment Intent ID"
          },
          "clientSecret": {
            "type": "string",
            "description": "Client secret for frontend"
          },
          "status": {
            "type": "string",
            "description": "Payment intent status"
          }
        },
        "required": [
          "success",
          "paymentId"
        ]
      },
      "CreatePaymentIntentRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "description": "Brand ID"
          },
          "campaignId": {
            "type": "string",
            "description": "Campaign ID"
          },
          "amountInCents": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Amount in cents"
          },
          "currency": {
            "type": "string",
            "default": "usd",
            "description": "Currency code"
          },
          "customerId": {
            "type": "string",
            "description": "Stripe Customer ID"
          },
          "description": {
            "type": "string",
            "description": "Payment description"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom metadata"
          }
        },
        "required": [
          "amountInCents"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/checkout/create": {
      "post": {
        "summary": "Create a Stripe checkout session",
        "description": "Creates a Stripe Checkout Session and records the payment in the database. Runs-service integration is BLOCKING.",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCheckoutSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/payment-intent/create": {
      "post": {
        "summary": "Create a Stripe payment intent",
        "description": "Creates a Stripe Payment Intent and records the payment in the database. Runs-service integration is BLOCKING.",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIntentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment intent created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentIntentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}